perf: agent form editor
