+++ packages/global/core/app/type.d.ts
@@ -93,7 +93,7 @@ export type AppDatasetSearchParamsType = {
   datasetSearchExtensionBg?: string;
 };
 
-export type AppSimpleEditFormType = {
+export type AppFormEditFormType = {
   // templateId: string;
   aiSettings: {
     [NodeInputKeyEnum.aiModel]: string;
