--- projects/app/src/pageComponents/app/detail/SimpleApp/styles.module.scss
@@ -1,10 +1,0 @@
-.EditAppBox {
-  &::-webkit-scrollbar-thumb {
-    background: #dfe2ea !important;
-    transition: background 1s;
-  }
-
-  &::-webkit-scrollbar-thumb:hover {
-    background: var(--chakra-colors-gray-300) !important;
-  }
-}
